# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: helmstudio platform API
  version: "1.0.0"
  description: >
    The local, single-machine API a studio's runtime SDK speaks, and the one
    the launcher speaks to the daemon. Served by the daemon at
    127.0.0.1:8700 and, for one studio with no daemon, by `helm dev` and the
    embedded provider (05 §2, §5a). The two are substitutable: one
    conformance suite runs against both (Q25). The in-process embedded
    provider has no token, and still enforces the studio's capabilities,
    read from its manifest, with the same 403 `capability_required`; the
    suite's capability cases run against both providers (M4 first review,
    #7).
servers:
  - url: http://127.0.0.1:8700/api/v1
...
security: []
paths:
...
  /studios/{id}/manifest:
    get:
      tags: [launcher]
      operationId: studiosManifest
      x-helm-group: studios
      x-helm-method: manifest
      summary: The resolved manifest text, with its digest, source and file.
      description: >
        Q17, Q18. The text as stored, byte for byte — the same bytes Export
        writes — so an editor round-trips comments and key order. `digest` is
        what `If-Match` on a save compares.
      parameters: [{ $ref: "#/components/parameters/StudioId" }]
      responses:
        "200":
          description: The manifest.
          content: { application/json: { schema: { $ref: "#/components/schemas/ManifestDoc" } } }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /launcher/manifests:validate:
    post:
      tags: [launcher]
      operationId: manifestsValidate
      x-helm-group: manifests
      x-helm-method: validate
      summary: Validate manifest or registry-entry text. Writes nothing.
      description: >
        Q17. The same validator `helm validate` and the daemon's loader use,
        through one bytes entry point — a test requires the same verdict and
        the same errors from all three for every file in
        `internal/manifest/testdata` and `studios/`.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ManifestText" } } }
      responses:
        "200":
          description: The verdict, whether or not it is valid.
          content: { application/json: { schema: { $ref: "#/components/schemas/ManifestCheck" } } }
        "400": { $ref: "#/components/responses/Invalid" }
  /launcher/manifests:edit:
    post:
      tags: [launcher]
      operationId: manifestsEdit
      x-helm-group: manifests
      x-helm-method: edit
      summary: Apply one field edit to manifest text, keeping comments and key order. Writes nothing.
      description: >
        Q17. The edit is a JSON pointer and a value; the daemon applies it to
        the YAML node tree and returns the new text. A `null` value removes the
        key. This is how the form and the YAML pane stay one document.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ManifestEdit" } } }
      responses:
        "200":
          description: The edited text, with the verdict for it.
          content: { application/json: { schema: { $ref: "#/components/schemas/ManifestCheck" } } }
        "400": { $ref: "#/components/responses/Invalid" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /launcher/manifests/{id}:
    put:
      tags: [launcher]
      operationId: manifestsSave
      x-helm-group: manifests
      x-helm-method: save
      summary: Save a local manifest or pointer, creating the Local entry that overrides lower sources.
      description: >
        R3a, Q5, Q18. Written to `<data>/studios/<id>.yaml`. An invalid
        document is refused with its errors (422) rather than saved broken.
        `If-Match` carries the digest the editor last read, so two tabs cannot
        silently overwrite each other; without it a save over an existing file
        is 428.

        Saving does not install, launch or rebuild anything. If the studio is
        installed, the card shows `rebuild_needed`, and the next install or
        launch goes through approval — which is what makes an Override that
        edits a command show the change instead of running it (Q10).
      parameters:
        - { $ref: "#/components/parameters/StudioId" }
        - { $ref: "#/components/parameters/IfMatch" }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ManifestText" } } }
      responses:
        "200":
          description: The saved manifest.
          content: { application/json: { schema: { $ref: "#/components/schemas/ManifestDoc" } } }
        "400": { $ref: "#/components/responses/Invalid" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422":
          description: The document is invalid, and was not saved.
          content: { application/json: { schema: { $ref: "#/components/schemas/ManifestCheck" } } }
        "428": { $ref: "#/components/responses/IfMatchRequired" }
    delete:
      tags: [launcher]
      operationId: manifestsRevert
      x-helm-group: manifests
      x-helm-method: revert
      summary: Revert a local manifest, so the source beneath it resolves again.
      description: >
        R3a, Q9. **The file is moved, never deleted**: it goes to
        `<data>/studios/_reverted/<id>-<timestamp>.yaml`, which is not scanned,
        and the response says where. A Revert that deleted could destroy the
        only copy of something the user wrote. The underscore matters — an id
        cannot start with one, so the directory can never collide with a
        studio's own.

        An installed or running studio is left alone: the group keeps running
        and the next install or launch goes through approval.
      parameters: [{ $ref: "#/components/parameters/StudioId" }]
      responses:
        "200":
          description: Where the file went, and what resolves now.
          content: { application/json: { schema: { $ref: "#/components/schemas/RevertResult" } } }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /launcher/manifests:import:
    post:
      tags: [launcher]
      operationId: manifestsImport
      x-helm-group: manifests
      x-helm-method: import
      summary: Report on manifests given as text or fetched from a URL, then add them on confirmation.
      description: >
        R3c, Q14, Q18. Two phases. Without `confirm`, every item is read and
        reported — what it is, whether it is valid, whether its id collides —
        and a `confirm` digest is returned; **nothing is written**. With
        `{confirm, resolutions}`, exactly the reported set is added, each item
        Overridden, Renamed or Skipped. A stale digest is 409 `preview_changed`.

        **Import is not install.** An imported manifest is a description until
        the user installs it, and installing it goes through approval like
        anything else (Q10) — including one with `local_path`, which the report
        flags as "builds in &lt;path&gt; on this Mac, without cloning".

        **A URL is fetched under Q14's rules:** `https://` only, no redirect to
        another scheme or host, refused when the host resolves to a loopback,
        link-local or private address, at most 1 MiB, 10 s, no cookies or
        credentials. Bytes that are not a manifest are never echoed back as
        text. Files, drops and paste are read by the page and sent as `text`:
        the daemon opens no path a request names.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ImportRequest" } } }
      responses:
        "200":
          description: The report, or what was added.
          content: { application/json: { schema: { $ref: "#/components/schemas/ImportReport" } } }
        "400": { $ref: "#/components/responses/Invalid" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /launcher/repositories:read:
    post:
      tags: [launcher]
      operationId: repositoriesRead
      x-helm-group: repositories
      x-helm-method: read
      summary: Read a repository's manifest without cloning or installing it. Writes nothing.
      description: >
        R3b, Q7, Q8, Q14. Git only, never a forge's HTTP API: a depth-1 fetch
        of the ref without file contents, then only `helmstudio.yaml` and
        `.gitmodules`, with the resolved commit recorded. A repository that
        ships no manifest answers with `found: false` and the resolved commit,
        which is what the editor pre-fills `repo` and `ref` from.

        `{path}` is "From folder": a **typed absolute path**, from which only
        `&lt;path&gt;/helmstudio.yaml` is read. There is deliberately no
        directory-listing endpoint, and a browser folder picker never gives a
        page an absolute path — so this is a thing a person types, not a thing
        a page discovers.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/RepositoryRead" } } }
      responses:
        "200":
          description: What the repository holds.
          content: { application/json: { schema: { $ref: "#/components/schemas/RepositoryResult" } } }
        "400": { $ref: "#/components/responses/Invalid" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "503": { $ref: "#/components/responses/Unavailable" }

  # ======================================================================
  # Launcher: the weights cache (01 §6, 02 §4; M3 as served)
  # ======================================================================
...
components:
  securitySchemes:
    studioToken:
      type: http
      scheme: bearer
      description: >
        HELM_TOKEN, minted per launch for one studio and its declared
        capabilities, revoked when its group stops (07 §3, Q6). Format
        `hs_live_` plus 52 base32 characters.
...
  parameters:
...
    IfMatch:
      name: If-Match
      in: header
      description: The etag last read. When given and stale, the write is refused with 409 `etag_mismatch`.
      schema: { type: string }
...
    StudioId:
      name: id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/StudioIdValue" }
...
  responses:
    Invalid:
      description: >
        The request is malformed. Codes: `bad_request`, `bad_filter`,
        `bad_cursor`, `invalid_document`.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    NotFound:
      description: Absent, deleted, or not the caller's to see. Code `not_found`.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    Conflict:
      description: >
        Codes: `etag_mismatch`, `name_taken`, `job_finished`,
        `already_running`, `not_running`, `port_conflict`, `heavy_conflict`
        (with `details.heavy`), `not_installed`, `conflict`, `in_use`,
        `missing`, `confirm_required` and `preview_changed` (with
        `details.preview`), `not_reclaimable`, and from M7:
        `approval_required` (with `details.approval`), `id_taken`,
        `not_fetched`, `already_exists`.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    IfMatchRequired:
      description: >
        The write needs `If-Match` and did not carry one. Saving a manifest
        over an existing file without naming the digest the editor read would
        let two tabs overwrite each other silently.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unavailable:
      description: >
        Something outside this machine could not be reached — a repository, a
        URL. Code `unavailable`, with what is known in `details`. Retrying may
        work, which is what separates it from the 4xx codes.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    Unprocessable:
      description: >
        Well-formed but refused. Codes: `outside_roots`, `cross_device`,
        `not_a_file`, `unknown_session`, `self_handoff`, `not_launchable`,
        `blocked`, `not_linkable`.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
...
    StudioIdValue:
      type: string
      pattern: "^[a-z][a-z0-9-]{1,38}[a-z0-9]$"
      description: A manifest id (schema/manifest.json `id`).
    JSONObject:
      type: object
      additionalProperties: true
    Error:
      type: object
      required: [error, message]
      properties:
        error: { type: string, description: "A stable snake_case code a program branches on." }
        message: { type: string, description: "A sentence for a person." }
        details: { type: object, additionalProperties: true }
...
    LibrarySource:
      type: string
      enum: [local, repo, registry, running]
      description: >
        Where this entry's manifest came from, highest first (Q6). `local` is
        `<data>/studios/<id>.yaml`; `repo` is `helmstudio.yaml` in the studio's
        own repository at the pinned ref, or in its installed checkout;
        `registry` is the inline manifest a pointer carries. `running` is a
        studio still running from a previous daemon whose manifest is gone.
...
    ManifestText:
      type: object
      additionalProperties: false
      required: [text]
      properties:
        text: { type: string, description: "The YAML, as written. A manifest or a registry entry; the validator tells them apart by shape." }
    ManifestDoc:
      type: object
      required: [id, text, digest, source, kind]
      properties:
        id: { $ref: "#/components/schemas/StudioIdValue" }
        text: { type: string, description: "Byte for byte what is stored — the same bytes Export writes." }
        digest: { type: string, description: "sha256 of the text. `If-Match` on a save compares this." }
        source: { $ref: "#/components/schemas/LibrarySource" }
        kind: { type: string, enum: [manifest, pointer] }
        file: { type: string, description: "Absolute path, when it is a file on this machine. The editor shows it and offers Reveal." }
        commit: { type: string, description: "The commit the text was read at, for a repository source." }
        reverted_to: { type: string, description: "Set on a Revert: where the file was moved to." }
    ManifestCheck:
      type: object
      description: "A verdict, whether or not it is valid. `document` is the parsed YAML as JSON, which is what the form renders from — the page never parses YAML itself (Q17)."
      required: [valid, kind, errors]
      properties:
        valid: { type: boolean }
        kind: { type: string, enum: [manifest, pointer, unknown] }
        errors: { type: array, items: { $ref: "#/components/schemas/ManifestError" } }
        warnings: { type: array, items: { $ref: "#/components/schemas/ManifestError" } }
        document: { $ref: "#/components/schemas/JSONObject" }
        text: { type: string, description: "The text this verdict is for; set by :edit, which returns what it wrote." }
        criteria: { $ref: "#/components/schemas/Criteria" }
    ManifestEdit:
      type: object
      additionalProperties: false
      required: [text, pointer]
      properties:
        text: { type: string }
        pointer: { type: string, description: "RFC 6901 JSON pointer to the field, e.g. /processes/0/cmd." }
        value:
          description: "The new value. Absent or null removes the key. Comments and key order around it are kept."
    RevertResult:
      type: object
      required: [id, moved_to]
      properties:
        id: { $ref: "#/components/schemas/StudioIdValue" }
        moved_to: { type: string, description: "`<data>/studios/_reverted/<id>-<timestamp>.yaml`. The file is never deleted (Q9)." }
        now: { $ref: "#/components/schemas/ManifestDoc" }
...
    ImportRequest:
      type: object
      additionalProperties: false
      description: "Without `confirm`, reports and writes nothing. With `{confirm, resolutions}`, adds exactly the reported set."
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/ImportItem" } }
        confirm: { type: string }
        resolutions: { type: array, items: { $ref: "#/components/schemas/ImportResolution" } }
...
    ImportReport:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/ImportItemReport" } }
        confirm: { type: string, description: "Send back to add exactly this set. Absent on the confirmed pass." }
    RepositoryRead:
      type: object
      additionalProperties: false
      description: "Exactly one of `{repo, ref}` or `{path}`. `path` is a typed absolute path, and only `<path>/helmstudio.yaml` is read."
      properties:
        repo: { type: string }
        ref: { type: string }
        path: { type: string }
    RepositoryResult:
      type: object
      required: [found]
      properties:
        found: { type: boolean, description: "False when the repository ships no helmstudio.yaml — which is the common case, and what the editor pre-fills from." }
        commit: { type: string }
        repo: { type: string }
        ref: { type: string }
        text: { type: string }
        check: { $ref: "#/components/schemas/ManifestCheck" }
        submodules: { type: array, items: { $ref: "#/components/schemas/Submodule" } }
...